+++ jest.config.js
@@ -1,5 +1,6 @@
 /** @type {import('ts-jest/dist/types').InitialOptionsTsJest} */
 module.exports = {
-  preset: "ts-jest",
-  testEnvironment: "node",
+  preset: "ts-jest/presets/js-with-babel",
+  testEnvironment: "jsdom",
+  transformIgnorePatterns: ["/node_modules/@jest"],
 };