+++ src/pokemonShowdownTypeHelper/types.ts
@@ -1,0 +1,1 @@
+export type TypesGroupedByDamageMultiplier = Map<number, string[]>;