+++ CHANGELOG.md
@@ -2,6 +2,42 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [0.0.7](https://github.com/coltonb/pokemon-showdown-type-helper/compare/v0.0.6...v0.0.7) (2023-06-16)
+
+
+### Bug Fixes
+
+* Benched pokemon tooltips not displaying ([#8](https://github.com/coltonb/pokemon-showdown-type-helper/issues/8)) ([d8b49c4](https://github.com/coltonb/pokemon-showdown-type-helper/commit/d8b49c49c581ee26e655184e8645ad976bfa2ca0))
+
+### [0.0.6](https://github.com/coltonb/pokemon-showdown-type-helper/compare/v0.0.5...v0.0.6) (2023-06-15)
+
+
+### Bug Fixes
+
+* Damage relation ordering and styling ([63664a2](https://github.com/coltonb/pokemon-showdown-type-helper/commit/63664a2d634dd3376de3ee371adcf2c657caa8a7))
+* Pokemon weaknesses should be correct when type is changed in battle.
+
+### [0.0.5](https://github.com/coltonb/pokemon-showdown-type-helper/compare/v0.0.4...v0.0.5) (2023-06-13)
+
+
+### Features
+
+* Firefox support ([0a6cedf](https://github.com/coltonb/pokemon-showdown-type-helper/commit/0a6cedfefd1ebb2eee4230cb431469f3d68222e2))
+
+### [0.0.4](https://github.com/coltonb/pokemon-showdown-type-helper/compare/v0.0.3...v0.0.4) (2023-06-06)
+
+
+### Features
+
+* Pokemon Stat Display ([#4](https://github.com/coltonb/pokemon-showdown-type-helper/issues/4)) ([647f4b1](https://github.com/coltonb/pokemon-showdown-type-helper/commit/647f4b1d209ff95f1946feeea83198d91b6548f4))
+
+### [0.0.3](https://github.com/coltonb/pokemon-showdown-type-helper/compare/v0.0.2...v0.0.3) (2023-03-29)
+
+
+### Bug Fixes
+
+* Tera type handling ([#1](https://github.com/coltonb/pokemon-showdown-type-helper/issues/1)) ([ea6adbc](https://github.com/coltonb/pokemon-showdown-type-helper/commit/ea6adbc6e4a33ad728652104f47f453fbdc6e979))
+
 ### [0.0.2](https://github.com/coltonb/pokemon-showdown-type-helper/compare/v0.0.1...v0.0.2) (2022-04-02)